+++ generators/base/support/options.mjs
