+++ generators/base/support/output/file-operations.mjs
