+++ generators/client/prompts.mjs
@@ -20,7 +20,7 @@ import chalk from 'chalk';
 import constants from '../generator-constants.cjs';
 import generatorDefaults from '../generator-defaults.cjs';
 import { clientFrameworkTypes } from '../../jdl/jhipster/index.mjs';
-import { httpsGet } from '../base/logic/index.mjs';
+import { httpsGet } from '../base/support/index.mjs';
 
 const { clientDefaultConfig } = generatorDefaults;
 const NO_CLIENT_FRAMEWORK = clientFrameworkTypes.NO;
