+++ generators/entity/support/asserts.mjs
