+++ generators/liquibase-changelogs/support/formatting.mjs
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { stringNullOrEmpty, textToArray, isSimpleText, htmlEncode } from '../../base/logic/index.mjs';
+import { stringNullOrEmpty, textToArray, isSimpleText, htmlEncode } from '../../base/support/index.mjs';
 
 /**
  * @private
